moved files from unnecessary directories
update includes for solver changes
