+++ framework/physics/solver.h
@@ -4,7 +4,7 @@
 #pragma once
 
 #include "framework/object.h"
-#include "framework/physics/basic_options/basic_options.h"
+#include "basic_options.h"
 #include "framework/parameters/parameter_block.h"
 #include <iostream>
 #include <utility>
