+++ lua/framework/physics/field_function/physics_get_field_func_list.cc
@@ -5,7 +5,7 @@
 #include "lua/framework/physics/field_function/field_functions.h"
 #include "lua/framework/console/console.h"
 #include "framework/runtime.h"
-#include "framework/physics/solver_base/solver.h"
+#include "framework/physics/solver.h"
 #include "framework/field_functions/field_function_grid_based.h"
 #include "framework/logging/log.h"
