+++ lua/framework/physics/solver/solver.cc
@@ -3,7 +3,7 @@
 
 #include "lua/framework/physics/solver/solver.h"
 #include "lua/framework/console/console.h"
-#include "framework/physics/solver_base/solver.h"
+#include "framework/physics/solver.h"
 #include "framework/field_functions/field_function_grid_based.h"
 #include "framework/event_system/physics_event_publisher.h"
 #include "framework/object_factory.h"
