+++ modules/diffusion/cfem_diffusion_solver.h
@@ -4,11 +4,6 @@
 #pragma once
 
 #include "modules/diffusion/diffusion_solver.h"
-#include "framework/mesh/mesh.h"
-#include "framework/physics/solver_base/solver.h"
-#include "framework/math/petsc_utils/petsc_utils.h"
-#include "framework/utils/timer.h"
-#include <map>
 
 namespace opensn
 {
