+++ modules/diffusion/fv_diffusion_solver.cc
@@ -2,13 +2,13 @@
 // SPDX-License-Identifier: MIT
 
 #include "modules/diffusion/fv_diffusion_solver.h"
+#include "framework/mesh/mesh_continuum/mesh_continuum.h"
+#include "framework/math/spatial_discretization/finite_volume/finite_volume.h"
+#include "framework/math/functions/scalar_spatial_material_function.h"
 #include "framework/runtime.h"
 #include "framework/object_factory.h"
 #include "framework/logging/log.h"
 #include "framework/utils/timer.h"
-#include "framework/mesh/mesh_continuum/mesh_continuum.h"
-#include "framework/math/spatial_discretization/finite_volume/finite_volume.h"
-#include "framework/math/functions/scalar_spatial_material_function.h"
 
 namespace opensn
 {
