+++ modules/linear_boltzmann_solvers/executors/nl_keigen.h
@@ -3,7 +3,7 @@
 
 #pragma once
 
-#include "framework/physics/solver_base/solver.h"
+#include "framework/physics/solver.h"
 #include "modules/linear_boltzmann_solvers/lbs_solver/lbs_solver.h"
 #include "modules/linear_boltzmann_solvers/lbs_solver/iterative_methods/nl_keigen_ags_solver.h"
 #include <petscsnes.h>
